chore: change method placement
